+++ .storybook/preview.ts
@@ -1,5 +1,4 @@
 import '../src/css/style.css';
-import '../src/css/palette-evertec.css';
 import DocumentationTemplate from './DocumentationTemplate.mdx';
 import { type Preview, setup } from '@storybook/vue3';
 import type { App } from 'vue';