+++ src/components/spartan/SInputDateBlock/index.ts
@@ -1,0 +1,1 @@
+export { default as SInputDateBlock } from './SInputDateBlock.vue';