+++ src/components/spartan/index.ts
@@ -41,3 +41,5 @@ export * from './SInputIncrementBlock';
 export * from './SSkeleton';
 export * from './SInputMask';
 export * from './SInputMaskBlock';
+export * from './SInputDate';
+export * from './SInputDateBlock';